Moving mobile logs with geolocation tests over to otlp_test.go
