+++ systemtest/ingest_test.go
@@ -28,17 +28,6 @@ import (
 	"os"
 	"strings"
 	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"github.com/tidwall/gjson"
-	"go.opentelemetry.io/collector/pdata/pcommon"
-	"go.opentelemetry.io/collector/pdata/plog"
-	"go.opentelemetry.io/collector/pdata/plog/plogotlp"
-	semconv "go.opentelemetry.io/collector/semconv/v1.5.0"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
 
 	"github.com/elastic/apm-server/systemtest"
 	"github.com/elastic/apm-server/systemtest/apmservertest"
@@ -46,6 +35,9 @@ import (
 	"github.com/elastic/apm-tools/pkg/approvaltest"
 	"github.com/elastic/apm-tools/pkg/espoll"
 	"github.com/elastic/go-elasticsearch/v8/esapi"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"github.com/tidwall/gjson"
 )
 
 func TestIngestPipeline(t *testing.T) {
@@ -306,61 +298,4 @@ func TestIngestPipelineBackwardCompatibility(t *testing.T) {
 		}, &indexResponse)
 		require.NoError(t, err)
 	}
-}
-
-func TestMobileLogsWithGeoLocation(t *testing.T) {
-	systemtest.CleanupElasticsearch(t)
-	srv := apmservertest.NewUnstartedServerTB(t)
-	err := srv.Start()
-	require.NoError(t, err)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	// Override local IP address to be found in "GeoLite2-City.mmdb".
-	md := metadata.New(map[string]string{"X-Forwarded-For": "195.55.79.118"})
-	ctx = metadata.NewOutgoingContext(ctx, md)
-
-	conn, err := grpc.Dial(serverAddr(srv), grpc.WithInsecure(), grpc.WithBlock(), grpc.WithDefaultCallOptions(grpc.UseCompressor("gzip")))
-	require.NoError(t, err)
-	defer conn.Close()
-
-	logsClient := plogotlp.NewGRPCClient(conn)
-
-	logs := newMobileLogs("a log message")
-	_, err = logsClient.Export(ctx, plogotlp.NewExportRequestFromLogs(logs))
-	require.NoError(t, err)
-
-	result := estest.ExpectDocs(t, systemtest.Elasticsearch, "logs-apm*", nil)
-	approvaltest.ApproveEvents(t, t.Name(), result.Hits.Hits)
-}
-
-func newMobileLogs(body interface{}) plog.Logs {
-	logs := plog.NewLogs()
-	resourceLogs := logs.ResourceLogs().AppendEmpty()
-	resourceAttrs := logs.ResourceLogs().At(0).Resource().Attributes()
-	resourceAttrs.PutStr(semconv.AttributeTelemetrySDKLanguage, "java")
-	resourceAttrs.PutStr(semconv.AttributeTelemetrySDKName, "android")
-
-	scopeLogs := resourceLogs.ScopeLogs().AppendEmpty()
-	otelLog := scopeLogs.LogRecords().AppendEmpty()
-	otelLog.SetTraceID(pcommon.TraceID{1})
-	otelLog.SetSpanID(pcommon.SpanID{2})
-	otelLog.SetSeverityNumber(plog.SeverityNumberInfo)
-	otelLog.SetSeverityText("Info")
-	otelLog.SetTimestamp(pcommon.NewTimestampFromTime(time.Unix(1, 0)))
-	otelLog.Attributes().PutStr("key", "value")
-	otelLog.Attributes().PutDouble("numeric_key", 1234)
-
-	switch b := body.(type) {
-	case string:
-		otelLog.Body().SetStr(b)
-	case int:
-		otelLog.Body().SetInt(int64(b))
-	case float64:
-		otelLog.Body().SetDouble(float64(b))
-	case bool:
-		otelLog.Body().SetBool(b)
-	}
-	return logs
 }
